hv: workaround the MOVDQA instruction emulation failure
This issue will cause certain board failed to boot WaaG, the direct
reason is that the hv failed to emulate MOVDQA instruction.

The hexdump of MOVDQA instruction is:
66 0F 7F 41 10 00 00 00 00 00 00 00 00 00 00

And due to the fact that this issue is hard to reproduce, so this
patch is not verified 100%, but the logic and the hexdump code are
correct and exact.

Signed-off-by: Xiaoguang Wu <xiaoguang.wu@intel.com>